refactor: refactor delete feature with react-redux
