+++ src/actions/actionDelTodo.js
@@ -1,3 +1,17 @@
+import { remove } from 'ramda'
+
 export const DEL_TODO = 'DEL_TODO'
 
-export const delTodo = todo_title => ({ type: DEL_TODO, payload: todo_title })
+const handleDelete = (index, todos) => {
+  const newTodos = remove(index, 1, todos)
+
+  return newTodos
+}
+
+export const delTodo = (selectedTodoKey, todos) => {
+  const newTodos = handleDelete(selectedTodoKey, todos)
+  return {
+    type: DEL_TODO,
+    payload: newTodos,
+  }
+}
